src/requirements.txt
apipkg==1.5
appdirs==1.4.4
asgiref==3.3.1
aspy.refactor-imports==2.1.1
attrs==20.3.0
black==20.8b1
cached-property==1.5.2
certifi==2020.11.8
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
curlylint==0.12.0
distlib==0.3.1
Django==3.1.3
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-livesync==0.5
django-simple-history==2.12.0
execnet==1.7.1
filelock==3.0.12
gunicorn==20.0.4
honeycomb-beeline==2.14.0
identify==1.5.9
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
isort==5.6.4
libhoney==1.10.0
more-itertools==8.6.0
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.4
parsy==1.3.0
pathspec==0.8.1
pathtools==0.1.2
pluggy==0.13.1
pre-commit==2.8.2
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytest-forked==1.3.0
pytest-xdist==2.1.0
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
regex==2020.10.28
requests==2.24.0
seed-isort-config==2.2.0
sentry-sdk==0.19.2
six==1.15.0
sqlparse==0.4.1
statsd==3.3.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.1.0
watchdog==0.10.3
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
